parsing cas response xml
added a simple testing file, will delete later, but anyway, this is
how libxmljs works, using real cas responses

modified lib/cas_validate to use the sax parser.  The sax parser is an
event parser, so it is annoying but fine.  anyway, what I do is parse
the returned doc on checking a ticket, and check for success or
failure.  On success, I log the person in and stash the user inthe
session.  I will also add the cas attributes when we turn that back on
for testing.

on failure, I just do nothing other than log it in the logs

Then at the end of parsing the cas doc, I redirect to the original
url, with the ticket removed from the query.  This will prevent double
tripping the ticket handler.