explanation/documentation for get_username, get_attributes
totally schizo on how to use these documentation comment blocks.  I
need to adopt some sort of documentation scheme.

Anyway, describe what each does.  If there is no info, return {} to the callback

I thought about migrating the old username functionality, but the
problem is that with connect, you can use any automated determination
of what response is acceptable, like you can with express.

On the other hand, there is no way to tell the user of this library
where I've stashed the username and the user attributes in the
session.  So it makes the most sense to just create accessor
functions, so that these can be retrieve with a minimum of fuss.

also added code so that while get_username just gets the username,
get_attributes will get the username plus the attributes.

No tests of this functionality yet.