worked out a test for ttl redis session stuff
ttl_test proves that the ST tickets are stored, and will disappear on
logout, and if you wait around longer than the set time to live.  The
test sets ttl to 2 seconds, and sets the server session ttl to 4
seconds (2*ttl), to demonstrate that allowing the session ticket in
redis to timeout sooner than the web server session defeats the
purpose of single sign out.