now parsing the other sample xml file
The other file passed attributes in a different way than the "usual"
way.  Looking into the CAS mailing list, passing attributes is ad hoc.
One way is to make an <attributes> element filled with <attribute>
elements.  Another way is to make a bunch of <attribute> elements, with
name='foo' value='bar' attributes for the <attribute>.