more progress
I think it is time to dig into what CAS is doing.  I see the ldap
server responding to the login request, but the "redirect" behavior is
not happening.  Instead the response is just nothing and error code
500

current way to set things up is codified in run docker.sh thing, as
usual.

run test is npx tap test/03<tab>

and it fails