futile attempt to update an ancient project
I don't use this anymore, and it seems my use of libxmljs 1 and 2 has
finally hit an end of life.  There are also gobs of libraries that are
obsolete and out of date used by libraries that I use.

Interesting how bitrot sets in when my code has not changed over the
past decade or so.